Day8: <DotNet> [Repository CRUD + Book CRUD Operations API Call Completed].

+++ dotNet/BookStore/BookStore.Api/Controllers/CategoryController.cs
@@ -3,11 +3,13 @@ using BookStore.Models.ViewModels;
 using BookStore.Repository;
 using Microsoft.AspNetCore.Mvc;
 using System.Linq;
+using System.Net;
 
 namespace BookStore.Api.Controllers
 {
     [Route("api/category")]
     [ApiController]
+    [ProducesResponseType(typeof(ListResponse<CategoryModel>), (int)HttpStatusCode.OK)]
     public class CategoryController : ControllerBase
     {
         CategoryRepository _categoryRepository = new CategoryRepository();
@@ -26,6 +28,8 @@ namespace BookStore.Api.Controllers
 
         [Route("{id}")]
         [HttpGet]
+        [ProducesResponseType(typeof(CategoryModel), (int)HttpStatusCode.OK)]
+
         public IActionResult GetCategory(int id)
         {
             var category = _categoryRepository.GetCategory(id);
@@ -35,8 +39,12 @@ namespace BookStore.Api.Controllers
 
         [Route("add")]
         [HttpPost]
+        [ProducesResponseType(typeof(CategoryModel), (int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(BadRequestObjectResult), (int)HttpStatusCode.BadRequest)]
         public IActionResult AddCategory(CategoryModel model)
         {
+            if (model == null)
+                return BadRequest("Model is null.");
             Category category = new Category()
             {
                 Id = model.Id,
@@ -49,8 +57,12 @@ namespace BookStore.Api.Controllers
 
         [Route("update")]
         [HttpPut]
+        [ProducesResponseType(typeof(CategoryModel), (int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(BadRequestObjectResult), (int)HttpStatusCode.BadRequest)]
         public IActionResult UpdateCategory(CategoryModel model)
         {
+            if (model==null)
+                return BadRequest("Model is null.");
             Category category = new Category()
             {
                 Id = model.Id,
@@ -63,8 +75,12 @@ namespace BookStore.Api.Controllers
 
         [Route("delete/{id}")]
         [HttpDelete]
+        [ProducesResponseType(typeof(bool), (int)HttpStatusCode.OK)]
+        [ProducesResponseType(typeof(BadRequestObjectResult), (int)HttpStatusCode.BadRequest)]
         public IActionResult DeleteCategory(int id)
         {
+            if (id==0)
+                return BadRequest("id is null.");
             var response = _categoryRepository.DeleteCategory(id);
             return Ok(response);
         }
